# apps/api/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/assistant/{id}": {
            "get": {
                "operationId": "AssistantController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            },
            "patch": {
                "operationId": "AssistantController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAssistantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            },
            "delete": {
                "operationId": "AssistantController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            }
        },
        "/assistant/user/{id}": {
            "get": {
                "operationId": "AssistantController_findByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            }
        },
        "/assistant": {
            "get": {
                "operationId": "AssistantController_findAll",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseAssistantDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            },
            "post": {
                "operationId": "AssistantController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAssistantDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            }
        },
...
        "/doctor/{id}": {
            "get": {
                "operationId": "DoctorController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            },
            "patch": {
                "operationId": "DoctorController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDoctorDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            },
            "delete": {
                "operationId": "DoctorController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseDoctorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
...
        "/availability/{id}": {
            "get": {
                "operationId": "AvailabilityController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            },
            "patch": {
                "operationId": "AvailabilityController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAvailabilityDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            },
            "delete": {
                "operationId": "AvailabilityController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            }
        },
        "/availability/doctor/{id}": {
            "get": {
                "operationId": "AvailabilityController_findByDoctor",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            }
        },
        "/availability": {
            "get": {
                "operationId": "AvailabilityController_findAll",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            },
            "post": {
                "operationId": "AvailabilityController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAvailabilityDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAvailabilityDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "availability"
                ]
            }
        },
...
    },
    "info": {
        "title": "MEDIGO API",
        "description": "API Endopoints for MEDIGO",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "ResponseUserPatientDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "email"
                ]
            },
...
            "ResponseSpecialityDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description"
                ]
            },
            "ResponseDoctorDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "speciality": {
                        "$ref": "#/components/schemas/ResponseSpecialityDto"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "phone",
                    "speciality",
                    "user"
                ]
            },
            "ResponseAssistantDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "doctor",
                    "user"
                ]
            },
            "CreateAssistantDto": {
                "type": "object",
                "properties": {
                    "doctorId": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    }
                },
                "required": [
                    "doctorId",
                    "userId"
                ]
            },
            "UpdateAssistantDto": {
                "type": "object",
                "properties": {
                    "doctorId": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    }
                }
            },
...
            "UpdateDoctorDto": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    },
                    "specialityId": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "number"
                    }
                }
            },
...
            "ResponseAvailabilityDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "start": {
                        "type": "string"
                    },
                    "end": {
                        "type": "string"
                    },
                    "day": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    }
                },
                "required": [
                    "id",
                    "start",
                    "end",
                    "day",
                    "available",
                    "doctor"
                ]
            },
            "CreateAvailabilityDto": {
                "type": "object",
                "properties": {
                    "start": {
                        "type": "string"
                    },
                    "end": {
                        "type": "string"
                    },
                    "day": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "doctorId": {
                        "type": "number"
                    }
                },
                "required": [
                    "start",
                    "end",
                    "day",
                    "available",
                    "doctorId"
                ]
            },
            "UpdateAvailabilityDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}